Attempt at adding player state info
This attempt was foiled by a logic issue: the TournamentState has a list of players and matches, and each match has a mini list of players. The match's list must be updated as well when a player update happens. The resolution to this issue is unlikely to be clean, unfortunately