Change miss counter to use Update() to send the player update
This is the wrong way to send updates, logically, but we just can't seem to nail down all the errors that occur when sending events while in the event. And errors which happen in the event handler will break the event overall, and can cause score discrepancies. Not worth the risk at this time.